refactor(core): 重构项目管理
 - 将 snapline 移出 tool 文件夹，因为它本身非 tool，放这里面会有误导
 - 优化 tool 文件夹，将 tool 内容移至 index 中，并更新项目中相关导入
